ocr/requirements.txt
fastapi
uvicorn
pymupdf
opencv-python-headless
numpy
pillow
python-multipart
requests
pandas
openpyxl